Refactor chat and memory instances
